Split out theme transfer actions
